+++ api/v1/class/benutzer/benutzer.arbeit.inc.php
@@ -90,6 +90,34 @@ namespace Arbeitszeit{
             }
         }
 
+        public static  function get_user_from_email($email){
+            $conn = new DB();
+            $sql = "SELECT * FROM `users` WHERE email = ?;";
+            $res = $conn->sendQuery($sql);
+            $res->execute([$email]);
+            if($res->rowCount() == 1){
+                $data = $res->fetch(\PDO::FETCH_ASSOC);
+                return $data;
+            } else {
+                Exceptions::error_rep("Could not find user with email '$email'.");
+                return false;
+            }
+        }
+
+        public function get_username_from_email($email){
+            $conn = new DB();
+            $sql = "SELECT username FROM `users` WHERE email = ?;";
+            $res = $conn->sendQuery($sql);
+            $res->execute([$email]);
+            if($res->rowCount() == 1){
+                $data = $res->fetch(\PDO::FETCH_ASSOC);
+                return $data["username"];
+            } else {
+                Exceptions::error_rep("Could not find user with email '$email'.");
+                return false;
+            }
+        }
+
         public function get_all_users(){
             $sql = "SELECT * FROM `users`;";
             $res = $this->db->sendQuery($sql);