+++ api/v1/toil/resources/addVacation.ep.toil.arbeit.inc.php
@@ -47,7 +47,7 @@ namespace Toil {
             }
         }
 
-        public function post()
+        public function post($post = null)
         {
 
         }