+++ api/v1/toil/resources/getLog.ep.toil.arbeit.inc.php
@@ -37,7 +37,7 @@ namespace Toil {
             }
         }
 
-        public function post()
+        public function post($post = null)
         {
 
         }