+++ api/v1/toil/resources/getUsers.ep.toil.arbeit.inc.php
@@ -41,7 +41,7 @@ namespace Toil {
             }
         }
 
-        public function post()
+        public function post($post = null)
         {
 
         }