+++ api/v1/toil/resources/getWorktimes.ep.toil.arbeit.inc.php
@@ -33,7 +33,7 @@ namespace Toil {
             }
         }
 
-        public function post()
+        public function post($post = null)
         {
 
         }