+++ suite/actions/worktime/add_vacation.php
@@ -2,9 +2,9 @@
 require $_SERVER["DOCUMENT_ROOT"] . "/api/v1/inc/arbeit.inc.php";
 use Arbeitszeit\Arbeitszeit;
 $worktime = new Arbeitszeit;
-$base_url = $arbeit->get_app_ini()["general"]["base_url"];
-$arbeit->auth()->login_validation();
-$vacation = $arbeit->vacation()->add_vacation($_POST["date-start"], $_POST["date-end"]);
+$base_url = $worktime->get_app_ini()["general"]["base_url"];
+$worktime->auth()->login_validation();
+$vacation = $worktime->vacation()->add_vacation(start: $_POST["date-start"], stop: $_POST["date-end"], username: $_SESSION["username"]);
 if(!$vacation){
     header("Location: http://{$base_url}/suite/?info=error_vacation");
 } else {